feat/separate_sidebar_component_into_different_files
