+++ src/atoms/IconAtom.js
@@ -1,0 +1,3 @@
+import { atom } from "jotai";
+
+export const iconsAtom = atom([]);
