+++ src/components/CreateComponent/index.jsx
@@ -6,13 +6,10 @@ import Sidebar from "./Sidebar";
 import { useAtom } from "jotai";
 import { canvasRefAtom } from "../../atoms/ComponentAtom";
 
-
 const CreateComponent = () => {
   const WrappedCanvasArea = forwardRef(CanvasArea);
   const canvasRef = useRef();
   const [, setCanvasAtom] = useAtom(canvasRefAtom);
-
-  
 
   useEffect(() => {
     setCanvasAtom(canvasRef);
